Refactor portions of nodeProjectGenerator into a more generic projectGenerator
