+++ packages/nx-plugin/src/generators/project/packageJson/files/package.json__template__
