+++ packages/nx-plugin/src/generators/project/packageJson/index.ts
@@ -1,1 +1,2 @@
+export * from "./copyPackageJsonTemplate";
 export * from "./standardizePackageJson";
